requirements.txt
aiogram==2.24
aiohttp==3.8.3
psycopg==3.1.8

aiofiles~=23.1.0
python-dotenv~=1.0.0